+++ coordinator/src/orderbook/db/mod.rs
@@ -1,2 +1,3 @@
 pub mod custom_types;
+pub mod matches;
 pub mod orders;